services/blockchain-service/requirements.txt
fastapi
uvicorn[standard]
web3
python-dotenv
pydantic
pydantic-settings
aiofiles
python-jose[cryptography]
passlib[bcrypt]
cryptography
redis
celery
py-solc-x
eth-account
pyjwt
requests